+++ src/qm_error_detector.c
@@ -3,14 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   qm_error_detector.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/07 00:22:19 by jvacaris          #+#    #+#             */
-/*   Updated: 2021/11/07 00:22:27 by jvacaris         ###   ########.fr       */
+/*   Updated: 2021/11/20 22:26:18 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+#define ERROR_OPEN_REDIR "minishel: syntax error near \
+unexpected token `%s'\n"
+#define LIT_NEWLINE "newline"
 
 /*
 * This function detects if there's quotation marks unclosed through the input 
@@ -37,8 +41,62 @@ int	qm_error_detector(char *str)
 	}
 	if (qm)
 	{
-		write(1, "minishell: unclosed quotation marks\n", 55);
+		write(1, "minishell: unclosed quotation marks\n", 37);
 		return (1);
 	}
+	return (0);
+}
+
+/**
+ * * Detects open redirection at the end of str return error if finds one 
+ * @param path	path
+*/
+int	has_last_redirection_open(const char *str)
+{
+	size_t	len;
+	int		ok;
+	char	*trim;
+
+	ok = 1;
+	trim = ft_strtrim(str, " ");
+	len = ft_strlen(trim);
+	if (ft_strchr("<>", trim[len - 1]) || !ft_strcmp(&trim[len - 2], ">|"))
+		ok = 0;
+	if (!ok)
+		printf(ERROR_OPEN_REDIR, LIT_NEWLINE);
+	free(trim);
+	return (ok);
+}
+
+/**
+ * * Detects open redirections 
+ * @param path	path
+*/
+int	has_redirection_open(const char **array)
+{
+	int		i;
+	char	*token;
+
+	i = 0;
+	token = NULL;
+	while (array[i])
+	{
+		if (ft_strcmp(array[i], ">") || ft_strcmp(array[i], "<") \
+		|| ft_strcmp(array[i], ">>") || ft_strcmp(array[i++], "<<"))
+		{
+			if (array[i])
+				token = ft_strdup(LIT_NEWLINE);
+			else if (ft_strcmp(array[i], ">") || ft_strcmp(array[i], "<") \
+				|| ft_strcmp(array[i], ">>") || ft_strcmp(array[i], "<<") \
+				|| ft_strcmp(array[i], "|"))
+				token = ft_strdup(array[i]);
+		}
+	}
+	if (token != NULL)
+	{
+		printf(ERROR_OPEN_REDIR, token);
+		free(token);
+		return (1);
+	}	
 	return (0);
 }